refactor: Move event pkg to ../storage/..
